+++ include/BoardManager.hpp
@@ -4,11 +4,20 @@
 #include "EventManager.hpp"
 #include "TextLoader.hpp"
 #include "../include/Events/TowerCreationEvent.hpp"
+#include "../include/Events/TowerRemoveEvent.hpp"
 #include <functional>
+#include <assert.h>
 
 using namespace std::placeholders;
 class BoardManager{
+
 private:
+    //value used to represnet an empty tile
+    int emptyTile = -1;
+
+    //int pair (used to assocaite a row and col)
+    typedef pair<int,int> intPair;
+
     //Store the textLoader to make requests for strings and constants
     shared_ptr<TextLoader> textLoader;
     //random map generator
@@ -31,6 +40,16 @@ private:
     // are the path to the exit
     vector<vector<int>> floorGrid;
 
+    /*
+     * Holds every obstacle on the above floor grid
+     * keyed by their type with a value of their position
+     */
+    unordered_map<int, intPair> allObstacles;
+
+    /*
+     * Vector to iterate through to explore each cardinal direction
+     */
+    vector<vector<int>> dirs = {{0,1},{0,-1},{1,0},{-1,0}};
 
 public:
   BoardManager(shared_ptr<EventManager> eventManager, shared_ptr<TextLoader> textLoader);
@@ -41,10 +60,20 @@ public:
   void newMap(MapChoices * newCustomization);
 
   void setMapBoards();
+  void setPathDistanceToReflectRealDistance();
+  void walkBackToEntries(int row, int col, int distanceSoFar,vector<vector<int>> traveled);
 
   void handleTowerCreation(const EventInterface& event);
+  void handleTowerRemove(const EventInterface& event);
 
+  bool hasMap();
+  bool isInMap(int row, int col);
+  bool isTower(int row, int col);
   bool isTowerOrObstacle(int row, int col);
+  bool isEntrance(int row, int col);
+  bool isExit(int row, int col);
+  bool isPath(int row, int col);
+  bool isObstacle(int row, int col);
 
   int getYDim();
   int getXDim();
@@ -69,7 +98,13 @@ public:
   vector<int>const & getFloorCol(int col);
   int getFloorRowCol(int row, int col);
 
-  vector<int> getEntryPositions();
+  vector<int>& getExitPositions();
+  vector<int>& getEntryPositions();
 
+  unordered_map<int,intPair>& getAllObstacles();
+
+private:
+  void addTower(int row, int col, int towerTypeNum);
+  void clearTowerOrObstacle(int row, int col);
 };
 #endif