+++ include/OptionsMenuScreen.hpp
@@ -18,7 +18,7 @@ private:
 public:
 
   OptionsMenuScreen(shared_ptr<EventManager> eventManager,shared_ptr<TextLoader> textLoader,int windowX, int windowY, int numItems, sf::Font font);
-
+  void initText();
   void moveUp();
   void moveDown();
   void draw(sf::RenderWindow &window);