+++ include/Towers/NormalFry.hpp
@@ -12,5 +12,6 @@ public:
 
   virtual void upgrade();
   virtual void setUpUnits();
+  void setProjectile(){}
 
 };