+++ include/Towers/SpicyFry.hpp
@@ -12,4 +12,6 @@ public:
 
   virtual void upgrade();
   virtual void setUpUnits();
+  void setProjectile(){}
+
 };