+++ src/Obstacles/TrashCan.cpp
@@ -1,7 +1,6 @@
 #include "../include/Obstacles/TrashCan.hpp"
-
-TrashCan::TrashCan(int row, int col, int price){
-  this -> row = row;
-  this -> col = col;
-  this -> price = price;
+TrashCan::TrashCan(shared_ptr<TextLoader> textLoader, string towerTypeID) : Obstacle(){
+  this -> towerTypeID = towerTypeID;
+  this -> textLoader = textLoader;
+  this -> price = textLoader -> getInteger(string("IDS_TC_PR"));
 }