+++ src/PlayingScreen.cpp
@@ -8,7 +8,36 @@ PlayingScreen::PlayingScreen(shared_ptr<EventManager> eventManager,shared_ptr<Te
   this -> textLoader = textLoader;
   this -> gameLogic = gameLogic;
   somethingChanged = true;
+  this -> registerDelegates();
 }
+
+/*
+ * Register the delegate method for this class
+ * with any events it needs to know about
+ */
+void PlayingScreen::registerDelegates(){
+  //bind our delegate function for tower creation events
+  EventManager::EventDelegate towerCreationEventDelegate = std::bind(&PlayingScreen::handleTowerCreation, this, _1);
+
+  //make an event and get its type
+  TowerCreationEvent towerCreationEvent = TowerCreationEvent();
+  EventType towerCreationEventType = towerCreationEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(towerCreationEventDelegate,
+  textLoader -> getString(string("IDS_PS_TC")),towerCreationEventType);
+
+
+  //bind our delegate function for tower remove events
+  EventManager::EventDelegate towerRemoveEventDelegate = std::bind(&PlayingScreen::handleTowerRemove, this, _1);
+
+  //make an event and get its type
+  TowerRemoveEvent towerRemoveEvent = TowerRemoveEvent();
+  EventType towerRemoveEventType = towerRemoveEvent.getEventType();
+  //register the delegate and its type
+  this -> eventManager -> registerDelegate(towerRemoveEventDelegate,
+  textLoader -> getString(string("IDS_PS_TR")),towerRemoveEventType);
+}
+
 
 void PlayingScreen::draw(sf::RenderWindow &window){
   //TODO change into real code for drawing the map!
@@ -27,6 +56,48 @@ void PlayingScreen::draw(sf::RenderWindow &window){
 
     somethingChanged = false;
   }
+
+}
+/*
+ * Handle any tower creation
+ * @param event: event of the tower creation
+ */
+void PlayingScreen::handleTowerCreation(const EventInterface& event){
+  /*
+   * cast the EventInterface reference to a CONST pointer to the
+   * TowerCreationEvent type which allows us to access variables and methods
+   * specific to TowerCreationEvent
+   */
+  const TowerCreationEvent* tcEvent = static_cast<const TowerCreationEvent*>(&event);
+  /*
+   * cast the "data" (a EventDataInterface) to a TowerCreationEventData type
+   * the .get() is because data is a unique_ptr and we need to grab the
+   * raw pointer inside of it for this
+   */
+  TowerCreationEventData* tcEventData = static_cast<TowerCreationEventData*>((tcEvent -> data).get());
+
+  somethingChanged = true;
+}
+
+/*
+ * Handle a tower remove event
+ * @param event: the tower remove event
+ */
+void PlayingScreen::handleTowerRemove(const EventInterface& event){
+  /*
+   * cast the EventInterface reference to a CONST pointer to the
+   * TowerRemoveEvent type which allows us to access variables and methods
+   * specific to TowerRemoveEvent
+   */
+  const TowerRemoveEvent* trEvent = static_cast<const TowerRemoveEvent*>(&event);
+  /*
+   * cast the "data" (a EventDataInterface) to a TowerRemoveEventData type
+   * the .get() is because data is a unique_ptr and we need to grab the
+   * raw pointer inside of it for this
+   */
+  TowerRemoveEventData* trEventData = static_cast<TowerRemoveEventData*>((trEvent -> data).get());
+
+  somethingChanged = true;
 
 }